gold-agent/requirements.txt
anthropic
gradio
yfinance
pandas
numpy
python-dotenv
requests
